test(ui): rename `userAuth.ts` to `factories.ts`
Centralizes API type factories in a single file so `mockTag`,
`mockNamespace`, and future factories land alongside `mockUserAuth`.
