+++ ui/apps/console/src/stores/__tests__/authStore.test.ts
@@ -2,7 +2,7 @@ import { describe, it, expect, vi, beforeEach, afterEach } from "vitest";
 import { useAuthStore } from "../authStore";
 import type { UserAuth } from "@/client";
 import { mockSdkResponse, type SdkResponse } from "@/tests/sdk";
-import { mockUserAuth } from "@/tests/userAuth";
+import { mockUserAuth } from "@/tests/factories";
 
 vi.mock("@/client", () => ({
   login: vi.fn(),
